+++ src/lib/converter/symbols.ts
@@ -1,9 +1,7 @@
 import assert from "assert";
 import ts from "typescript";
 import {
-    Comment,
     DeclarationReflection,
-    DocumentReflection,
     IntrinsicType,
     LiteralType,
     ReferenceReflection,
@@ -28,9 +26,6 @@ import {
 import { convertJsDocAlias, convertJsDocCallback } from "./jsdoc";
 import { getHeritageTypes } from "./utils/nodes";
 import { removeUndefined } from "./utils/reflections";
-import { basename, dirname, resolve } from "path";
-import { MinimalSourceFile, readFile } from "../utils";
-import { nicePath } from "../utils/paths";
 
 const symbolConverters: {
     [K in ts.SymbolFlags]?: (
@@ -297,41 +292,7 @@ function convertNamespace(
             exportSymbol,
         );
         context.finalizeDeclarationReflection(reflection);
-
-        let relativeTo = reflection.comment?.sourcePath;
-        if (relativeTo) {
-            relativeTo = dirname(relativeTo);
-            const tags = reflection.comment?.getTags("@document") || [];
-            reflection.comment?.removeTags("@document");
-            for (const tag of tags) {
-                const path = Comment.combineDisplayParts(tag.content);
-
-                let file: MinimalSourceFile;
-                try {
-                    const resolved = resolve(relativeTo, path);
-                    file = new MinimalSourceFile(readFile(resolved), resolved);
-                } catch {
-                    context.logger.warn(
-                        context.logger.i18n.failed_to_read_0_when_processing_document_tag_in_1(
-                            nicePath(path),
-                            nicePath(reflection.comment!.sourcePath!),
-                        ),
-                    );
-                    continue;
-                }
-
-                const { content, frontmatter } =
-                    context.converter.parseRawComment(file);
-                const docRefl = new DocumentReflection(
-                    basename(file.fileName).replace(/\.[^.]+$/, ""),
-                    reflection,
-                    content,
-                    frontmatter,
-                );
-                reflection.addChild(docRefl);
-                context.project.registerReflection(docRefl);
-            }
-        }
+        context.converter.processDocumentTags(reflection);
     }
 
     convertSymbols(